backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
django-filter==23.5
Pillow==10.1.0
python-decouple==3.8
requests==2.31.0
geoip2==4.8.0
django-extensions==3.2.3
redis==5.0.1
django-redis==5.4.0
django-cacheops==7.2
django-cloudinary-storage==0.3.0
cloudinary==1.35.0
drf-spectacular==0.27.1 